Split e2e helpers by suite
